fix: delete

+++ src/main/java/memo/backend/service/NoteService.java
@@ -1,19 +1,27 @@
 package memo.backend.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import memo.backend.model.Note;
 import memo.backend.model.User;
 import memo.backend.repository.NoteRepository;
+import memo.backend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class NoteService {
     @Autowired
     private NoteRepository noteRepository;
+
+    @Autowired
+    private UserRepository userRepository;
 
     public Note createNote(Note note) {
         return noteRepository.save(note);
@@ -47,25 +55,27 @@ public class NoteService {
         return noteRepository.findById(noteId).orElseThrow(() -> new RuntimeException("Nota não encontrada"));
     }
 
-    public void deleteNote(UUID noteId) {
-        noteRepository.deleteById(noteId);
+    @Transactional
+    public void deleteNote(UUID noteId, UUID userId) {
+        Optional<Note> noteOptional = noteRepository.findById(noteId);
+        Optional<User> userOptional = userRepository.findById(userId);
+
+        if (noteOptional.isPresent() && userOptional.isPresent()) {
+            User user = userOptional.get();
+            Note note = noteOptional.get();
+
+            user.getNotes().remove(note);
+            userRepository.save(user);
+            note.setOwner(null);
+            note.setUsers(null); // Remove todas as referências de usuários
+            noteRepository.save(note); // Salva as alterações antes de deletar
+            noteRepository.deleteById(noteId); // Deleta a nota
+        } else {
+            throw new EntityNotFoundException("Nota não encontrada com o ID: " + noteId);
+        }
     }
 
     public List<Note> listAllByUser(UUID userId) {
         return noteRepository.findAllByUser(userId);
-    }
-
-    @Transactional
-    public void deleteOrUnlinkNotesFromUser(User user, List<UUID> notesIds) {
-        List<Note> notes = noteRepository.findAllById(notesIds);
-
-        for (Note note : notes) {
-            if (note.getOwner().equals(user)) {
-                noteRepository.delete(note);
-            } else {
-                note.getUsers().remove(user);
-                noteRepository.save(note);
-            }
-        }
     }
 }
